requirements.txt
textual-dev==1.5.1
art==6.2